Renamed applications -> quick-links
