+++ src/components/structure/header/navigation.tsx
@@ -1,12 +1,12 @@
 "use client";
 
 import UserNav from "./user-nav";
-import ApplicationSwitcher from "./application-switcher";
+import QuickLinkDropdown from "./quick-link-dropdown";
 
 export default function Navigation({ noAuth }: { noAuth?: boolean }) {
   return (
     <nav className="w-full bg-background py-4 flex justify-between px-2">
-      <ApplicationSwitcher noAuth={noAuth} />
+      <QuickLinkDropdown noAuth={noAuth} />
       <UserNav noAuth={noAuth} />
     </nav>
   );
